+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.2 (unreleased)
 
 - Fixed error with FFI 1.13.0+
+- Added friendly graph optimization levels
 
 ## 0.3.1 (2020-05-18)